+++ extensions/open-in-sublime-text/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Open in Sublime Text Changelog
 
+## [New Command]
+
+- Open the clipboard content in Sublime Text.
+
 ## [Initial Version] - 2023-12-31
 
 - Open current Finder directory, selected folder or file in Sublime Text.